Move ActiveRecord::Delegation tests to a separate test case.
